pyproject.toml
[project]
name = "flink"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "docker>=7.1.0",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "kafka-python>=2.3.0",
    "marimo>=0.20.4",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pgcli>=4.4.0",
    "polars>=1.38.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "pyflink>=1.0",
]
